fix: solves unwanted unmount of wrapped components
When withResponsiveProps invokes React.createElement, the first is
withTheme(ResponsiveProps), this is always a new component. React cleans
up this component when props change and mounts it again. Visually it is
hard to notice the error, as very few DOM changes happen,depending on
the setting. However this is noticeable with components running effects
or life cycles.